Update modules to VSF 1.11
