--- store/state.ts
@@ -1,5 +1,0 @@
-import { BraintreeState } from '../types/BraintreeState'
-
-export const state: BraintreeState = {
-    trans: null
-}
