+++ types/BraintreeState.ts
@@ -1,5 +1,5 @@
 // This object should represent structure of your modules Vuex state
 // It's a good practice is to name this interface accordingly to the KET (for example mailchimpState)
 export interface BraintreeState {
-    trans: string[]
+  trans: string[]
 }
